requirements.txt
apache-tvm==0.10.0
black==23.1.0
click==8.1.3
flake8==6.0.0
isort==5.12.0
onnx==1.13.0
onnxoptimizer==0.3.6
Pillow==9.4.0
pydantic==1.10.4
pytest==7.2.1
xgboost==1.7.3
